requirements.txt
SpeechRecognition
pyttsx3
pyaudio
scikit-learn
numpy
pandas


# ----- Core Backend -----
flask==3.0.3
google-generativeai==0.7.2
wikipedia-api==0.7.1
requests==2.31.0

# ----- Machine Learning -----
scikit-learn==1.3.2
numpy==1.26.4
scipy==1.11.4

# ----- Desktop App -----
pillow==10.2.0
opencv-python==4.7.0.72
pyttsx3==2.90
SpeechRecognition==3.10.0
pyaudio==0.2.13

# ----- EXE Packaging -----
pyinstaller==6.3.0
